Módulo Perfis

+++ app/Http/Controllers/Admin/PlanController.php
@@ -87,12 +87,21 @@ class PlanController extends Controller
     public function destroy($url)
     {
 
-        $plan = $this->repository->where("url", $url)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where("url", $url)
+            ->first();
 
         if (!$plan)
             return redirect()->back();
 
-        $plan->delete();
+        if ($plan->details()->count() > 0) {
+            return redirect()->back()->with("error", "Existem detalhes vinculados a esse plano, portanto não pode deletar");
+        } else {
+            $plan->delete();
+        }
+
+
 
         return redirect()->route("plans.index");
     }
